frontend/apps/ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.24.3
python-dotenv==1.0.0
